backend/requirements.txt
asgiref==3.9.1
channels==4.3.1
dj_database_url==3.0.1
Django==5.2.7
daphne==4.2.1
decorator==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-picklefield==3.3
django-redis==6.0.0
djangorestframework==3.16.1
django-storages==1.14.6
jieba==0.42.1
numpy==2.3.3
ollama==0.6.0
openai==2.2.0
pgvector==0.4.1
pika==1.3.2
python-dotenv==1.1.1
Requests==2.32.5
celery==5.5.3
cachetools==5.5.2
channels==4.3.1
channels_redis==4.3.0
fsspec==2024.6.1
gevent==25.9.1
pydantic==2.11.7
pydantic_core==2.33.2
redis==5.2.1
whitenoise==6.9.0
psycopg2-binary==2.9.10